% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{getParameters.data.table}
\alias{getParameters.data.table}
\title{Get parameters from data table}
\usage{
\method{getParameters}{data.table}(x, ...)
}
\arguments{
\item{x}{data table objects}

\item{...}{additional arguments currently not used}
}
\value{
parameter vector
}
\description{
Get parameters from data table
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
