% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{normL2_indiv}
\alias{normL2_indiv}
\title{Fast normL2}
\usage{
normL2_indiv(
  data,
  prd0,
  errmodel = NULL,
  est.grid,
  fix.grid,
  times = NULL,
  attr.name = "data",
  mycores = 1
)
}
\arguments{
\item{data}{datalist object}

\item{prd0}{base prediction function}

\item{errmodel}{error model}

\item{est.grid}{Data frame with columns \code{ID}, \code{condition}, and columns for the
parameters}

\item{fix.grid}{Data frame with columns \code{ID}, \code{condition}, and columns for the
parameters}

\item{times}{Optional vector of timepoints where \code{prd0} should be evaluated}

\item{attr.name}{Character, which attribute to asign the obj value for combination with other obj functions}

\item{mycores}{You can use mycores > 1 for parallel computation of conditions to speed up single fits or profiles. Use carefully when also performing multi-start optimization via mstrust.}
}
\value{
objective function
}
\description{
Creates the objective function for a population of subjects with the same underlying model
and parametrization.
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
