% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{P_indiv}
\alias{P_indiv}
\title{Generate population parameter transformation from base transformation and individual grids}
\usage{
P_indiv(p0, est.grid, fix.grid)
}
\arguments{
\item{p0}{Base parameter transformation function}

\item{est.grid}{Data frame with columns \code{ID}, \code{condition}, and columns for the
parameters}

\item{fix.grid}{Data frame with columns \code{ID}, \code{condition}, and columns for the
parameters}
}
\value{
Parameter transformation function
}
\description{
Generate population parameter transformation from base transformation and individual grids
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
