% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{PRD_indiv}
\alias{PRD_indiv}
\title{Population function based on common base prediction function and grid of individuals}
\usage{
PRD_indiv(prd0, est.grid, fix.grid)
}
\arguments{
\item{prd0}{Prediction function}

\item{est.grid}{Data frame with columns \code{ID}, \code{condition}, and columns for the
parameters}

\item{fix.grid}{Data frame with columns \code{ID}, \code{condition}, and columns for the
parameters}
}
\value{
Prediction function
}
\description{
Population function based on common base prediction function and grid of individuals
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
