% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{renameDerivPars}
\alias{renameDerivPars}
\title{Update attr(prediction, "deriv") to correct par.grid.outer names}
\usage{
renameDerivPars(pred0, pars, est.grid, cn)
}
\arguments{
\item{pred0}{\code{prd0(times,pars)[[1]]}}

\item{pars}{pars}

\item{est.grid}{Data frame with columns \code{ID}, \code{condition}, and columns for the
parameters}

\item{cn}{name of condition}
}
\value{
pred0 with updated deriv argument
}
\description{
Update attr(prediction, "deriv") to correct par.grid.outer names
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
