% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{timepointL2_indiv}
\alias{timepointL2_indiv}
\title{DatapointL2 without env access}
\usage{
timepointL2_indiv(
  name,
  time,
  value,
  sigma = 1,
  attr.name = "timepointL2",
  condition,
  prd_indiv
)
}
\arguments{
\item{name}{character, state of observable name}

\item{time}{numeric, time point}

\item{value}{numeric, value}

\item{sigma}{numeric, uncertainty}

\item{attr.name}{character, the attribute name by which the value will be visible
for combination of several objetive functions}

\item{condition}{character, the condition}

\item{prd_indiv}{population prediction function}
}
\value{
objective function
}
\description{
DatapointL2 without env access
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
