% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIndiv.R
\name{unclass_parvec}
\alias{unclass_parvec}
\title{Unclass a parvec}
\usage{
unclass_parvec(x)
}
\arguments{
\item{x}{object of class parvec}
}
\value{
Named numeric with the same elements as x
}
\description{
Usual unclass, but keep names
}
